refactor: extract THonoEnv from route.ts
